♻️ 동아리 필터를 단일 반환하도록 변경
