+++ src/main/kotlin/team/gsm/flooding/domain/club/controller/ClubController.kt
@@ -6,9 +6,11 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import team.gsm.flooding.domain.club.dto.request.CreateClubRequest
 import team.gsm.flooding.domain.club.dto.response.FindClubFilterResponse
+import team.gsm.flooding.domain.club.entity.ClubType
 import team.gsm.flooding.domain.club.usecase.CreateClubUsecase
 import team.gsm.flooding.domain.club.usecase.FindClubFilterUsecase
 
@@ -27,8 +29,10 @@ class ClubController(
 		}
 
 	@GetMapping
-	fun findClubFilter(): ResponseEntity<FindClubFilterResponse> =
-		findClubFilterUsecase.execute().run {
+	fun findClubFilter(
+		@RequestParam type: ClubType,
+	): ResponseEntity<FindClubFilterResponse> =
+		findClubFilterUsecase.execute(type).run {
 			ResponseEntity.ok(this)
 		}
 }
