+++ src/main/kotlin/team/gsm/flooding/domain/club/usecase/FindClubFilterUsecase.kt
@@ -2,8 +2,9 @@ package team.gsm.flooding.domain.club.usecase
 
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import team.gsm.flooding.domain.club.dto.response.ClubFilterResponse
 import team.gsm.flooding.domain.club.dto.response.FindClubFilterResponse
-import team.gsm.flooding.domain.club.entity.ClubStatus
+import team.gsm.flooding.domain.club.entity.ClubType
 import team.gsm.flooding.domain.club.repository.ClubRepository
 import team.gsm.flooding.global.util.UserUtil
 
@@ -13,20 +14,10 @@ class FindClubFilterUsecase(
 	private val clubRepository: ClubRepository,
 	private val userUtil: UserUtil,
 ) {
-	fun execute(): FindClubFilterResponse {
-		val clubs = clubRepository.findAll()
+	fun execute(type: ClubType): FindClubFilterResponse {
+		val clubs = clubRepository.findByType(type)
 		val currentUser = userUtil.getUser()
-		val myClubs =
-			clubs.filter { club ->
-				club.leader == currentUser
-			}
-		val otherClubs =
-			clubs
-				.filter { club ->
-					(club.leader != currentUser) && (club.status != ClubStatus.PENDING)
-				}
-				.groupBy { it.type }
 
-		return FindClubFilterResponse.toDto(myClubs, otherClubs)
+		return FindClubFilterResponse(clubs.map { ClubFilterResponse.toDto(it, currentUser) })
 	}
 }
